chore: consolidate bin resolution tests into build.spec.ts
